Refactor dmnpmml to dmn.pmml
