--- dmn-pmml-quarkus-example/src/test/java/org/kie/dmnpmml/kogito/quarkus/example/DMNMiningModelTest.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.dmnpmml.kogito.quarkus.example;
-
-import org.junit.jupiter.api.Test;
-
-import io.quarkus.test.junit.QuarkusTest;
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import io.restassured.path.json.config.JsonPathConfig;
-
-import static io.restassured.RestAssured.given;
-import static io.restassured.config.JsonConfig.jsonConfig;
-import static org.hamcrest.Matchers.*;
-
-@QuarkusTest
-public class DMNMiningModelTest {
-
-    static {
-        RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
-    }
-
-    @Test
-    public void testEvaluateMiningModelDMN() {
-        String inputData = "{\"input1\":200.0, \"input2\":-1.0, \"input3\":2.0}";
-        given()
-                .config(RestAssured.config().jsonConfig(jsonConfig().numberReturnType(JsonPathConfig.NumberReturnType.DOUBLE)))
-                .contentType(ContentType.JSON)
-                .body(inputData)
-                .when()
-                .post("/TestMiningModelDMN")
-                .then()
-                .statusCode(200)
-                .body("SumMiningModelBKM", is("function SumMiningModelBKM( input1, input2, input3 )"))
-                .body("input1", is(comparesEqualTo(200))) // was input
-                .body("input2", is(comparesEqualTo(-1))) // was input
-                .body("input3", is(comparesEqualTo(2))) // was input
-                .body("Decision", is(comparesEqualTo(-299))) // real decision output
-        ;
-    }
-}
