--- dmn-pmml-quarkus-example/src/test/java/org/kie/dmnpmml/kogito/quarkus/example/DecisionTreeTest.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.dmnpmml.kogito.quarkus.example;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.junit.jupiter.api.Test;
-
-import io.quarkus.test.junit.QuarkusTest;
-import io.restassured.RestAssured;
-
-import static org.kie.dmnpmml.kogito.quarkus.example.CommonTestUtils.testDescriptive;
-import static org.kie.dmnpmml.kogito.quarkus.example.CommonTestUtils.testResult;
-
-@QuarkusTest
-class DecisionTreeTest {
-
-    private static final String BASE_PATH = "/DecisionTree";
-    private static final String TARGET = "decision";
-
-    static {
-        RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
-    }
-
-    @Test
-    void testEvaluateDecisionTreeResult() {
-        String inputData = "{\"temperature\":30.0, \"humidity\":10.0}";
-        testResult(inputData, BASE_PATH, TARGET, "sunglasses");
-    }
-
-    @Test
-    void testEvaluateDecisionTreeDescriptive() {
-        String inputData = "{\"temperature\":30.0, \"humidity\":10.0}";
-        final Map<String, Object> expectedResultMap = new HashMap<>();
-        expectedResultMap.put(TARGET, "sunglasses");
-        expectedResultMap.put("weatherdecision", "sunglasses");
-        testDescriptive(inputData, BASE_PATH, TARGET, expectedResultMap);
-    }
-
-}
